==> 0_Select_absence.ipynb <==
import pandas as pd
from tqdm import tqdm
import numpy as np
import random
import pymysql
from mlpckg.util_method import *

#DB에서 불러오기
db = pymysql.connect(host='localhost'
                    ,port = 3306
                    ,user='root'
                    ,password='23359584'
                    ,db='ladybug_project'
                    ,charset='utf8') 

cursor = db.cursor() 
sql = "SELECT * FROM all_coccinellid"

df_all = pd.read_sql_query(sql, db)

df_all[[ 'Source', 'State', 'Year', 'Species', 'Latitude',
       'Longitude', 'public_positional_accuracy']].to_csv("_all_coccinellid_Preprocessed.csv")


db.close()

df_ab = df_all[(df_all.Species=='Harmonia axyridis')|(df_all.Species=='Coccinella septempunctata')].reset_index(drop=True)

df_ab_struct = df_ab[(df_ab.Source != 'I-naturalist')&
       (df_ab.Source != 'The Lost Ladybugs')&
       (df_ab.Source !='Bugguide.net')].reset_index(drop=True)




df_presence = df_all[(df_all.Species=='Coccinella novemnotata')|(df_all.Species=='Coccinella transversoguttata')|
                     (df_all.Species=='Hippodamia parenthesis')|(df_all.Species=='Adalia bipunctata')]

df_presence.to_csv("./presence_only.csv")

### There are two ways to select absence data points. You can choose the way you like it.

#Method 1. select absence, We thought that about 10 percent of all absence data points were appropriate
random_sampled_absence = df_ab.iloc[random.sample(list(df_ab.index), k=int(len(df_ab)/10))].sort_values(by=[]).reset_index(drop=True)

random_sampled_absence

#Method 2. This method randomly extracts absence in the state to be proportional to the frequency of appearance in the state of the entire presence data. 
# To avoid clumping, pull out the indexes of rows aligned by coordinate at regular intervals.
it = ['Coccinella novemnotata',
'Coccinella transversoguttata',
'Hippodamia parenthesis',
'Adalia bipunctata']

for name in it:
    df_tmp = df_presence[df_presence.Species==name].reset_index(drop=True)
    #비율 측정
    ratio = 3 # Larger ratio returns more absence data points
    N=int(len(df_tmp)*ratio) 
    df_state = (df_tmp.State.value_counts()/len(df_tmp))*N

    #비율에 따라 분모 x 생성
    state_x ={}
    for i in df_state.index:
        if i in df_ab.State.unique():
            if int(len(df_ab[df_ab.State==i].index) / df_state[df_state.index==i][0]) != 0:
                state_x[i]=int(len(df_ab[df_ab.State==i].index) / df_state[df_state.index==i][0])
            else:
                state_x[i] = 1 



    #             
    many_idx = []
    for i in tqdm(df_state.index):
        #필요한 숫자가 후보군 전체보다 클 때
        if (len(df_ab[df_ab.State==i].index)) < df_state[i]:
                for j in range(len(df_ab[df_ab.State==i].index)):
                    if (j%(state_x[i])) == 0:
                        many_idx.append(df_ab[df_ab.State==i].index[j])
        else:
            for j in range(len(df_ab[df_ab.State==i].index)):
                if (j%(state_x[i]+1)) == 0:
                    many_idx.append(df_ab[df_ab.State==i].index[j])


    #타겟종의 absence후보들 선정완료된 dataframe생성
    df_abs = df_ab.iloc[many_idx].reset_index(drop=True)
      
    
    # 만약 구조적데이터의 개수가 presence가 더 많다면!
    if len(df_tmp[(df_tmp.Source != 'I-naturalist')&
       (df_tmp.Source != 'The Lost Ladybugs')&
       (df_tmp.Source !='Bugguide.net')]) > \
        len(df_abs[(df_abs.Source != 'I-naturalist')&
        (df_abs.Source != 'The Lost Ladybugs')&
        (df_abs.Source !='Bugguide.net')]):
        index=[]
        for i in df_tmp.State.unique():
            if len(df_ab_struct[df_ab_struct.State==i].index) <= 4:
                for j in range(len(df_ab_struct[df_ab_struct.State==i].index)):
                        index.append(df_ab_struct[df_ab_struct.State==i].index[j])
            else:
                if len(df_ab_struct[df_ab_struct.State==i].index) <6:
                    for t in (random.sample(list(df_ab_struct[df_ab_struct.State==i].index), k=4)):
                        index.append(t)
                else:
                    for t in (random.sample(list(df_ab_struct[df_ab_struct.State==i].index), k=6)):
                        index.append(t)


        tlqkf = df_ab_struct.iloc[index].reset_index(drop=True)


            
    
    #final dataframe is here. ex) novem_fin , trans_fin...
    globals()['{}_fin'.format(name.split(" ")[1][:5])] = pd.concat([df_tmp,df_abs,tlqkf])

#number of absence point 
for name in it:
    print(len(globals()['{}_fin'.format(name.split(" ")[1][:5])][globals()['{}_fin'.format(name.split(" ")[1][:5])].Species!=name]))

#choose what you want and to_csv.

#_____.to_csv("./absence_only.csv")
