==> 2_scaling.ipynb <==
import pandas as pd
from sklearn.preprocessing import MinMaxScaler
from tqdm import tqdm

#call your test dataframe which includes Absence and Presence datapoint of your all target Species
df=pd.read_csv("./Data/2nd_manipulated_data/20220807_final_test_all.csv")

#In my case, these are abandoned Species
df = df[(df.Species!='Brachiacantha ursina')&(df.Species!='Coccinella trifasciata')].reset_index(drop=True)

#Outlier drop
index_drop = []
for i in tqdm(df.iloc[:,9:].columns):
    q3 = df[i].quantile(0.75) 
    q1 = df[i].quantile(0.25)
    iqr = q3 - q1
    for j in df[i]:
        if j in index_drop:
            continue
        elif str(df.iloc[j].Year) not in i:
            continue

        data = df[i][j]

        if data > q3 + 1.5 * iqr or data < q1 - 1.5 * iqr:
            index_drop.append(j)


df = df.drop(index_drop).reset_index(drop=True)
            

#minmax scaling is depend on target Species. so if you need, modify this list
it = ['Adalia bipunctata' , 'Coccinella novemnotata',
 'Coccinella transversoguttata' , 'Hippodamia parenthesis' ]



for i in it:
    df_copy = df.copy()
    target_species=i
    df_copy = df_copy[(df_copy.Species==target_species)|(df_copy.Species=='Coccinella septempunctata')|(df_copy.Species=='Harmonia axyridis')].reset_index(drop=True)
    print(df_copy.Species.unique())
    
    scaler = MinMaxScaler()
    #check this index location, it must be only columns named by species name
    scaler.fit(df_copy.iloc[:,9:])

    # scaled dataframe
    df_tmp = pd.DataFrame(scaler.transform(df_copy.iloc[:,9:]), columns=df_copy.iloc[:,9:].columns, index=list(df_copy.index.values))

    # concat with original columns that they are not columns named by species name
    df_fin = pd.concat([df_copy.iloc[:,:8],df_tmp],axis=1).reset_index(drop=True)


    df_fin.to_csv("./Data/2nd_manipulated_data/20220808_"+target_species.split()[1][:5]+"_test_minmax.csv", index=False)
