==> 6_AOO_crawling.ipynb <==
from mlpckg.util_method import *
import urllib.request
from bs4 import BeautifulSoup
import time
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys

# select_key() argument ("1.novem/ 2.trans/ 3.bipun/ 4.paren ")
# select_key() return (target_species,key)

num = 2
target_species,key= select_key(num)[0], select_key(num)[1]

#insert your path for AOO Crawling
smooth_path = "C:/Users/DAEGYEONGLEE/Desktop/ladybug_project/Data/3rd_manipulated_data/forAOO/"

nav = {}
ex = []
ex2 = []
ex3 = []
ex4 = []

for i in range(2007,2022):
    try:
        driver = webdriver.Chrome()
        driver.get('http://geocat.kew.org/')
        driver.find_element(By.XPATH,'/html/body/section[2]/div/header/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'//*[@id="welcome"]/div[2]/ul/li[2]/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/a').click()
        time.sleep(0.5)


        # send_keys argument:absolute path of AOO csv
        driver.find_element(By.CSS_SELECTOR,"#add_source_container > div > div > div > ul > li.upload_source.selected > div > div.idle > div > form > input[type=file]").send_keys(smooth_path + target_species+"_aoo"+str(i)+".csv")


        time.sleep(15)
        # upload your file
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/div/div[3]/a[2]').send_keys(Keys.ENTER)
        time.sleep(20)
        #check aoo text
        driver.find_element(By.XPATH,'//*[@id="toggle_analysis"]/span').click()
        time.sleep(0.5)
        # save aoo text
        nav[str(i)]=int(driver.find_element(By.CSS_SELECTOR,'#tools > div > div.center > div.center-tool > div.right > div.analysis > div > div.analysis_data > ul > li:nth-child(2) > p:nth-child(2)').text.replace(",","").split(".")[0])
        # quit
        driver.quit()
        print(i)
    except Exception as e:
        ex.append(i)
        driver.quit()
        print(i,"Error", e)
        pass


for i in ex:
    try:
        driver = webdriver.Chrome()
        driver.get('http://geocat.kew.org/')
        driver.find_element(By.XPATH,'/html/body/section[2]/div/header/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'//*[@id="welcome"]/div[2]/ul/li[2]/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/a').click()
        time.sleep(0.5)


        # send_keys 뒤에 절대경로 넣기
        driver.find_element(By.CSS_SELECTOR,"#add_source_container > div > div > div > ul > li.upload_source.selected > div > div.idle > div > form > input[type=file]").send_keys(smooth_path + target_species+"_aoo"+str(i)+".csv")


        time.sleep(15)
        #업로드시키기
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/div/div[3]/a[2]').send_keys(Keys.ENTER)
        time.sleep(20)
        #aoo 텍스트 따오기
        driver.find_element(By.XPATH,'//*[@id="toggle_analysis"]/span').click()
        time.sleep(0.5)
        #aoo 텍스트 저장
        nav[str(i)]=int(driver.find_element(By.CSS_SELECTOR,'#tools > div > div.center > div.center-tool > div.right > div.analysis > div > div.analysis_data > ul > li:nth-child(2) > p:nth-child(2)').text.replace(",","").split(".")[0])
        #창닫기
        driver.quit()
        print(i)
    except Exception as e:
        ex2.append(i)
        driver.quit()
        print(i,"Error", e)
        pass



for i in ex2:
    try:
        driver = webdriver.Chrome()
        driver.get('http://geocat.kew.org/')
        driver.find_element(By.XPATH,'/html/body/section[2]/div/header/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'//*[@id="welcome"]/div[2]/ul/li[2]/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/a').click()
        time.sleep(0.5)


        # send_keys 뒤에 절대경로 넣기
        driver.find_element(By.CSS_SELECTOR,"#add_source_container > div > div > div > ul > li.upload_source.selected > div > div.idle > div > form > input[type=file]").send_keys(smooth_path + target_species+"_aoo"+str(i)+".csv")


        time.sleep(15)
        #업로드시키기
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/div/div[3]/a[2]').send_keys(Keys.ENTER)
        time.sleep(20)
        #aoo 텍스트 따오기
        driver.find_element(By.XPATH,'//*[@id="toggle_analysis"]/span').click()
        time.sleep(0.5)
        #aoo 텍스트 저장
        nav[str(i)]=int(driver.find_element(By.CSS_SELECTOR,'#tools > div > div.center > div.center-tool > div.right > div.analysis > div > div.analysis_data > ul > li:nth-child(2) > p:nth-child(2)').text.replace(",","").split(".")[0])
        #창닫기
        driver.quit()
        print(i)
    except Exception as e:
        ex3.append(i)
        driver.quit()
        print(i,"Error", e)
        pass


for i in ex3:
    try:
        driver = webdriver.Chrome()
        driver.get('http://geocat.kew.org/')
        driver.find_element(By.XPATH,'/html/body/section[2]/div/header/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'//*[@id="welcome"]/div[2]/ul/li[2]/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/a').click()
        time.sleep(0.5)


        # send_keys 뒤에 절대경로 넣기
        driver.find_element(By.CSS_SELECTOR,"#add_source_container > div > div > div > ul > li.upload_source.selected > div > div.idle > div > form > input[type=file]").send_keys(smooth_path + target_species+"_aoo"+str(i)+".csv")


        time.sleep(15)
        #업로드시키기
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/div/div[3]/a[2]').send_keys(Keys.ENTER)
        time.sleep(20)
        #aoo 텍스트 따오기
        driver.find_element(By.XPATH,'//*[@id="toggle_analysis"]/span').click()
        time.sleep(0.5)
        #aoo 텍스트 저장
        nav[str(i)]=int(driver.find_element(By.CSS_SELECTOR,'#tools > div > div.center > div.center-tool > div.right > div.analysis > div > div.analysis_data > ul > li:nth-child(2) > p:nth-child(2)').text.replace(",","").split(".")[0])
        #창닫기
        driver.quit()
        print(i)
    except Exception as e:
        ex.append(i)
        driver.quit()
        print(i,"Error", e)
        pass


for i in ex4:
    try:
        driver = webdriver.Chrome()
        driver.get('http://geocat.kew.org/')
        driver.find_element(By.XPATH,'/html/body/section[2]/div/header/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'//*[@id="welcome"]/div[2]/ul/li[2]/a').click()
        time.sleep(0.5)
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/a').click()
        time.sleep(0.5)


        # send_keys 뒤에 절대경로 넣기
        driver.find_element(By.CSS_SELECTOR,"#add_source_container > div > div > div > ul > li.upload_source.selected > div > div.idle > div > form > input[type=file]").send_keys(smooth_path + target_species+"_aoo"+str(i)+".csv")


        time.sleep(15)
        #업로드시키기
        driver.find_element(By.XPATH,'/html/body/div[1]/div[1]/div/div[3]/div[2]/div/div/div/div/ul/li[6]/div/div[3]/a[2]').send_keys(Keys.ENTER)
        time.sleep(20)
        #aoo 텍스트 따오기
        driver.find_element(By.XPATH,'//*[@id="toggle_analysis"]/span').click()
        time.sleep(0.5)
        #aoo 텍스트 저장
        nav[str(i)]=int(driver.find_element(By.CSS_SELECTOR,'#tools > div > div.center > div.center-tool > div.right > div.analysis > div > div.analysis_data > ul > li:nth-child(2) > p:nth-child(2)').text.replace(",","").split(".")[0])
        #창닫기
        driver.quit()
        print(i)
    except Exception as e:
        driver.quit()
        print(i,"Finally Error", e)
        pass


nav = dict(sorted(nav.items(), key=lambda x: x[0]))



AOO_fin = pd.DataFrame({"Year":nav.keys(),"AOO":nav.values()})
AOO_fin['Year'] = AOO_fin['Year'].astype(float)
AOO_fin['AOO'] = AOO_fin['AOO'].astype(float)

############# reduction rate
Reduction_rate = pd.DataFrame({"Start_Year":[],"AOO":[]})
Reduction_rate['Start_Year'] = list(range(2007,2013))

try:
    Reduction_rate['AOO'][0] = (float(AOO_fin['AOO'][0]) - float(AOO_fin['AOO'][9]))/float(AOO_fin['AOO'][0])
    Reduction_rate['AOO'][1] = (float(AOO_fin['AOO'][1]) - float(AOO_fin['AOO'][10]))/float(AOO_fin['AOO'][1])
    Reduction_rate['AOO'][2] = (float(AOO_fin['AOO'][2]) - float(AOO_fin['AOO'][11]))/float(AOO_fin['AOO'][2])
    Reduction_rate['AOO'][3] = (float(AOO_fin['AOO'][3]) - float(AOO_fin['AOO'][12]))/float(AOO_fin['AOO'][3])
    Reduction_rate['AOO'][4] = (float(AOO_fin['AOO'][4]) - float(AOO_fin['AOO'][13]))/float(AOO_fin['AOO'][4])
    Reduction_rate['AOO'][5] = (float(AOO_fin['AOO'][5]) - float(AOO_fin['AOO'][14]))/float(AOO_fin['AOO'][5])

except:
    pass

Reduction_rate['AOO'].max()

Reduction_rate.to_csv('./Data/3rd_manipulated_data/AOO_and_Reduction_Rate/' + target_species + '_reduction_rate_' + str(key) + '_in_test.csv')
AOO_fin.to_csv('./Data/3rd_manipulated_data/AOO_and_Reduction_Rate/' + target_species + '_AOO_' + str(key) + '_in_test.csv')
